+++ libe/drivers/i2c/sht31.h
@@ -29,7 +29,7 @@ extern "C" {
 #define SHT31_REPEATABILITY_LOW         0x16
 
 
-int8_t sht31_open(struct i2c_device *dev, struct i2c_master *master, float reference, int32_t resolution);
+int8_t sht31_open(struct i2c_device *dev, struct i2c_master *master, int32_t reference, int32_t resolution);
 #define sht31_close(dev) i2c_close(dev)
 int8_t sht31_heater(struct i2c_device *dev, bool on);
 int8_t sht31_read(struct i2c_device *dev, float *t, float *h);
